Update image degrade script for arch linux
I had to move away from ubuntu 24.04 to get good wayland desktop perf after switching to nvidia. While adjusting the font list, I found that the fonts I thought to be imagemagick builtins are actually duplicates (aliases) of other fonts and arch doesn't have the aliases, so the duplicate fonts are removed. The source packages of the fonts are also better documented.